Add return and else
Hoisted Tinder Yields to engine so the engine can be informed about jumps in the script.  Jump now updates the last line visited, and Return will return to the last stored jump allowing for basic sub routines.  Else is added to the goto which causes a jump when reached by a means other than a jump allowing more streamlined jump logic.