tinder updates
Break and Continue for loops, set from and forwhile.  Additionally, key value pairs can now take strings as well as identifiers.  At was removed as it was redundant and didn't make any sense.  Write can now omit write if the first value on a line is a string, but still allows a follow on expression, call can now omit call.